requirements.txt
bintrees
lxml
matplotlib
numpy
pandas
pillow
pymzml
pycocotools
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.1